test: inline getopts parser tests
