+++ src/arg_group.rs
@@ -16,7 +16,7 @@ impl NamedArgGroup {
     }
 }
 
-#[derive(Debug, Deserialize, Clone)]
+#[derive(Debug, Deserialize, Clone, PartialEq)]
 #[serde(deny_unknown_fields)]
 #[serde(rename_all = "kebab-case")]
 pub struct ArgGroup {
